refactor(state): separate migrations and backups
Move core schema SQL into an embedded migration file, extract migration execution and backup retention into focused modules, and ignore non-PV backup files during migration backup discovery.
